fix blessed
