+++ blessed/blessed-tests.ts
@@ -1,0 +1,1 @@
+import * as blessed from 'blessed'
